+++ rules/mock/failing.go
@@ -79,6 +79,11 @@ func (*failingService) OnSign(_ context.Context, _ *rules.ReqMetadata, _ *rules.
 	return rules.FAILED
 }
 
+// OnSignVoluntaryExit is called when a request to sign a voluntary exit needs to be approved.
+func (*failingService) OnSignVoluntaryExit(_ context.Context, _ *rules.ReqMetadata, _ *rules.SignData) rules.Result {
+	return rules.FAILED
+}
+
 // ExportSlashingProtection exports the slashing protection data.
 func (*failingService) ExportSlashingProtection(_ context.Context) (map[[48]byte]*rules.SlashingProtection, error) {
 	return make(map[[48]byte]*rules.SlashingProtection), nil